Extract LogResults as separate module
