+++ test/cli_test.exs
@@ -64,8 +64,4 @@ defmodule CliTest do
     assert parsed_rslt.max  == "55.803"
     assert parsed_rslt.stddev == "2.043"
   end
-
-  test "write_log to console" do
-    assert write_log(parse_result({@test_ping1, %{}, :std_out})) == :ok
-  end
 end
